# agenthub-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent-profiles": {
      "get": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "List Agent Profiles Api",
        "operationId": "list_agent_profiles_api_api_v1_agent_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AgentProfileOut__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "Create Agent Profile Api",
        "operationId": "create_agent_profile_api_api_v1_agent_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AgentProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-profiles/{profile_id}/files": {
      "get": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "List Profile Files Api",
        "operationId": "list_profile_files_api_api_v1_agent_profiles__profile_id__files_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_str__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-profiles/{profile_id}/files/{filename}": {
      "get": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "Get Profile File Api",
        "operationId": "get_profile_file_api_api_v1_agent_profiles__profile_id__files__filename__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileContentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "Update Profile File Api",
        "operationId": "update_profile_file_api_api_v1_agent_profiles__profile_id__files__filename__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceFileWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileContentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-profiles/{profile_id}/file-toggles": {
      "get": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "Get Profile Toggles Api",
        "operationId": "get_profile_toggles_api_api_v1_agent_profiles__profile_id__file_toggles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileTogglesOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agent-profiles"
        ],
        "summary": "Update Profile Toggles Api",
        "operationId": "update_profile_toggles_api_api_v1_agent_profiles__profile_id__file_toggles_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "boolean"
                },
                "default": {},
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileTogglesOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent-instances/{agent_instance_id}/workspace/files": {
      "get": {
        "tags": [
          "agent-instances"
        ],
        "summary": "List Workspace Files Api",
        "operationId": "list_workspace_files_api_api_v1_agent_instances__agent_instance_id__workspace_files_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileListOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-instances/{agent_instance_id}/workspace/files/{filename}": {
      "get": {
        "tags": [
          "agent-instances"
        ],
        "summary": "Read Workspace File Api",
        "operationId": "read_workspace_file_api_api_v1_agent_instances__agent_instance_id__workspace_files__filename__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Instance Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileContentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agent-instances"
        ],
        "summary": "Write Workspace File Api",
        "operationId": "write_workspace_file_api_api_v1_agent_instances__agent_instance_id__workspace_files__filename__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Instance Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceFileWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WorkspaceFileContentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/members/agents": {
      "post": {
        "tags": [
          "members"
        ],
        "summary": "Create Agent Member Api",
        "operationId": "create_agent_member_api_api_v1_members_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentMemberCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemberOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentMemberCreateRequest": {
        "properties": {
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "agent_instance_id": {
            "type": "string",
            "title": "Agent Instance Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "display_name",
          "agent_instance_id"
        ],
        "title": "AgentMemberCreateRequest"
      },
      "AgentProfileCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Role"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "soul_md": {
            "type": "string",
            "minLength": 1,
            "title": "Soul Md"
          },
          "agents_md": {
            "type": "string",
            "title": "Agents Md",
            "default": ""
          },
          "profile_md": {
            "type": "string",
            "title": "Profile Md",
            "default": ""
          },
          "bootstrap_md": {
            "type": "string",
            "title": "Bootstrap Md",
            "default": ""
          },
          "memory_md": {
            "type": "string",
            "title": "Memory Md",
            "default": ""
          },
          "heartbeat_md": {
            "type": "string",
            "title": "Heartbeat Md",
            "default": ""
          },
          "enabled_files_json": {
            "type": "string",
            "title": "Enabled Files Json",
            "default": "{}"
          },
          "model_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Model Name",
            "default": "gpt-4.1-mini"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "top_p": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Top P",
            "default": 1.0
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "planning_mode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Mode"
          },
          "is_active": {
            "type": "integer",
            "title": "Is Active",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "soul_md"
        ],
        "title": "AgentProfileCreateRequest"
      },
      "AgentProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "soul_md": {
            "type": "string",
            "title": "Soul Md"
          },
          "agents_md": {
            "type": "string",
            "title": "Agents Md"
          },
          "profile_md": {
            "type": "string",
            "title": "Profile Md"
          },
          "bootstrap_md": {
            "type": "string",
            "title": "Bootstrap Md"
          },
          "memory_md": {
            "type": "string",
            "title": "Memory Md"
          },
          "heartbeat_md": {
            "type": "string",
            "title": "Heartbeat Md"
          },
          "enabled_files_json": {
            "type": "string",
            "title": "Enabled Files Json"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature"
          },
          "top_p": {
            "type": "number",
            "title": "Top P"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "planning_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Mode"
          },
          "is_active": {
            "type": "integer",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "description",
          "soul_md",
          "agents_md",
          "profile_md",
          "bootstrap_md",
          "memory_md",
          "heartbeat_md",
          "enabled_files_json",
          "model_name",
          "temperature",
          "top_p",
          "max_output_tokens",
          "reasoning_effort",
          "planning_mode",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "AgentProfileOut"
      },
...
      "ApiResponse_AgentProfileOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/AgentProfileOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[AgentProfileOut]"
      },
...
      "ApiResponse_MemberOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/MemberOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[MemberOut]"
      },
...
      "ApiResponse_WorkspaceFileContentOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/WorkspaceFileContentOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[WorkspaceFileContentOut]"
      },
      "ApiResponse_WorkspaceFileListOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/WorkspaceFileListOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[WorkspaceFileListOut]"
      },
      "ApiResponse_WorkspaceFileTogglesOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/WorkspaceFileTogglesOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[WorkspaceFileTogglesOut]"
      },
...
      "ApiResponse_list_AgentProfileOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentProfileOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[AgentProfileOut]]"
      },
...
      "ApiResponse_list_str__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[str]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceFileContentOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "name",
          "content"
        ],
        "title": "WorkspaceFileContentOut"
      },
      "WorkspaceFileInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "name",
          "exists",
          "size",
          "enabled"
        ],
        "title": "WorkspaceFileInfo"
      },
      "WorkspaceFileListOut": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceFileInfo"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "WorkspaceFileListOut"
      },
      "WorkspaceFileTogglesOut": {
        "properties": {
          "enabled_files": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Enabled Files"
          }
        },
        "type": "object",
        "required": [
          "enabled_files"
        ],
        "title": "WorkspaceFileTogglesOut"
      },
...
      "WorkspaceFileWriteRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 500000,
            "title": "Content",
            "default": ""
          }
        },
        "type": "object",
        "title": "WorkspaceFileWriteRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}